requirements.txt
pytest
requests
python-dotenv
marshmallow
assertpy
